Feat: Meteorites now added to the chart.
Locations of meteorite landings on earth are now added to the chart.
Function plotPoints was added to handle that.
The function is called from within d3.json which drew the earth to ensure that the points are drawn over earth.
Changed the color of earth from random coloring to #A3CCFF.
Circles representing meteorite landing have random colors from a set of 8 colors.
The size of each circle depends on the mass of each meteorite it represents.