Feat: Added mousemove function to move tooltip with the pointer.
Up to this point when the tooltip is displayed, it doesn't move if the pointer moves.
The tooltip jumps if the hovered circle changes.
The mousover function changes the position of the tooltip when the pointer moves.
Now the tooltip will always follow the tooltip.
Added tooltipStatus variable to execute the mouseover commands only when the tooltip is displayed 'tooltipStatus = true'.